asteroid_mining_homework1_and_homework2/users/api/requirements.txt
Flask==3.0.2 
psycopg-pool==3.2.1
psycopg-binary==3.1.18
psycopg==3.1.18
PyJWT==2.8.0
